requirements.txt
PyCrypto==2.5
pil==1.1.7
python-twitpic
tweepy
